chore: places the visualization and corresponding tests into their own directories
