+++ cognee/__init__.py
@@ -5,8 +5,8 @@ from .api.v1.datasets.datasets import datasets
 from .api.v1.prune import prune
 from .api.v1.search import SearchType, get_search_history, search
 from .api.v1.visualize import visualize_graph
-from cognee.infrastructure.visualization.cognee_network_visualization import (
-    create_cognee_style_network_with_logo,
+from cognee.modules.visualization.cognee_network_visualization import (
+    cognee_network_visualization,
 )
 
 # Pipelines
