+++ cognee/api/v1/visualize/visualize.py
@@ -1,5 +1,5 @@
-from cognee.infrastructure.visualization.cognee_network_visualization import (
-    create_cognee_style_network_with_logo,
+from cognee.modules.visualization.cognee_network_visualization import (
+    cognee_network_visualization,
 )
 from cognee.infrastructure.databases.graph import get_graph_engine
 import logging
@@ -14,7 +14,7 @@ async def visualize_graph():
     graph_data = await graph_engine.get_graph_data()
     logging.info(graph_data)
 
-    graph = await create_cognee_style_network_with_logo(graph_data)
+    graph = await cognee_network_visualization(graph_data)
     logging.info("The HTML file has been stored on your home directory! Navigate there with cd ~")
 
     return graph
